requirements/docs.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe --output-file=requirements/docs.txt requirements/docs.in
#
aabbtree==2.6.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
alabaster==0.7.12
    # via
    #   -c requirements/dev.txt
    #   sphinx
attrs==20.3.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
    #   pinttrs
babel==2.9.0
    # via
    #   -c requirements/dev.txt
    #   sphinx
beautifulsoup4==4.9.3
    # via
    #   -c requirements/dev.txt
    #   pydata-sphinx-theme
cerberus==1.3.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
certifi==2020.12.5
    # via
    #   -c requirements/dev.txt
    #   requests
cftime==1.4.1
    # via
    #   -c requirements/dev.txt
    #   netcdf4
chardet==4.0.0
    # via
    #   -c requirements/dev.txt
    #   requests
click==7.1.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
cycler==0.10.0
    # via
    #   -c requirements/dev.txt
    #   matplotlib
dask==2021.3.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
docutils==0.16
    # via
    #   -c requirements/dev.txt
    #   pybtex-docutils
    #   sphinx
    #   sphinx-panels
    #   sphinxcontrib-bibtex
iapws==1.5.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
idna==2.10
    # via
    #   -c requirements/dev.txt
    #   requests
imagesize==1.2.0
    # via
    #   -c requirements/dev.txt
    #   sphinx
importlib-metadata==3.7.2
    # via
    #   -c requirements/dev.txt
    #   pint
jinja2==2.11.3
    # via
    #   -c requirements/dev.txt
    #   sphinx
kiwisolver==1.3.1
    # via
    #   -c requirements/dev.txt
    #   matplotlib
latexcodec==2.0.1
    # via
    #   -c requirements/dev.txt
    #   pybtex
markupsafe==1.1.1
    # via
    #   -c requirements/dev.txt
    #   jinja2
matplotlib==3.3.4
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
mock==4.0.3
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
netcdf4==1.5.6
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
numpy==1.20.1
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
    #   cftime
    #   matplotlib
    #   netcdf4
    #   pandas
    #   scipy
    #   xarray
packaging==20.9
    # via
    #   -c requirements/dev.txt
    #   pint
    #   sphinx
pandas==1.2.3
    # via
    #   -c requirements/dev.txt
    #   xarray
pillow==8.1.2
    # via
    #   -c requirements/dev.txt
    #   matplotlib
pint==0.16.1
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
    #   pinttrs
pinttrs==21.1.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
pybtex-docutils==1.0.0
    # via
    #   -c requirements/dev.txt
    #   sphinxcontrib-bibtex
pybtex==0.24.0
    # via
    #   -c requirements/dev.txt
    #   pybtex-docutils
    #   sphinxcontrib-bibtex
pydata-sphinx-theme==0.5.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
pygments==2.8.1
    # via
    #   -c requirements/dev.txt
    #   sphinx
pyparsing==2.4.7
    # via
    #   -c requirements/dev.txt
    #   matplotlib
    #   packaging
python-dateutil==2.8.1
    # via
    #   -c requirements/dev.txt
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   -c requirements/dev.txt
    #   babel
    #   pandas
pyyaml==5.4.1
    # via
    #   -c requirements/dev.txt
    #   dask
    #   pybtex
requests==2.25.1
    # via
    #   -c requirements/dev.txt
    #   sphinx
ruamel.yaml.clib==0.2.2
    # via
    #   -c requirements/dev.txt
    #   ruamel.yaml
ruamel.yaml==0.16.13
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
scipy==1.6.1
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
    #   iapws
six==1.15.0
    # via
    #   -c requirements/dev.txt
    #   cycler
    #   latexcodec
    #   pybtex
    #   python-dateutil
snowballstemmer==2.1.0
    # via
    #   -c requirements/dev.txt
    #   sphinx
soupsieve==2.2
    # via
    #   -c requirements/dev.txt
    #   beautifulsoup4
sphinx-copybutton==0.3.1
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
sphinx-gallery==0.8.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
sphinx-panels==0.5.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
sphinx==3.5.2
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
    #   pydata-sphinx-theme
    #   sphinx-copybutton
    #   sphinx-gallery
    #   sphinx-panels
    #   sphinxcontrib-bibtex
sphinxcontrib-applehelp==1.0.2
    # via
    #   -c requirements/dev.txt
    #   sphinx
sphinxcontrib-bibtex==2.2.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
sphinxcontrib-devhelp==1.0.2
    # via
    #   -c requirements/dev.txt
    #   sphinx
sphinxcontrib-htmlhelp==1.0.3
    # via
    #   -c requirements/dev.txt
    #   sphinx
sphinxcontrib-jsmath==1.0.1
    # via
    #   -c requirements/dev.txt
    #   sphinx
sphinxcontrib-qthelp==1.0.3
    # via
    #   -c requirements/dev.txt
    #   sphinx
sphinxcontrib-serializinghtml==1.1.4
    # via
    #   -c requirements/dev.txt
    #   sphinx
tinydb==4.4.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/docs.in
tqdm==4.59.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
typing-extensions==3.7.4.3
    # via
    #   -c requirements/dev.txt
    #   importlib-metadata
urllib3==1.26.3
    # via
    #   -c requirements/dev.txt
    #   requests
xarray==0.17.0
    # via
    #   -c requirements/dev.txt
    #   -r requirements/main.in
zipp==3.4.1
    # via
    #   -c requirements/dev.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==54.1.1
    # via
    #   -c requirements/dev.txt
    #   cerberus
    #   sphinx
    #   xarray
